Add card views to stories

+++ app/src/main/java/com/rrajath/orange/data/Item.java
@@ -23,5 +23,5 @@ public class Item {
     public long score;
     public String title;
     ArrayList<Long> parts;
-    public long descendants;
+    public int descendants;
 }
